2_9: OTA gui size u32 (dong bo logic OTA voi 2.13)
0xA0 doi setUint16 -> setUint32 tai buf+2: firmware >64KB lam u16 tran
-> xoa thieu sector -> hong bank. Firmware 2.9 cu doc u16 van dung vi
2 byte cao = 0 khi size <64KB. Di kem firmware 0xA529000F (doc u32 +
commit-last hoan ghi trang dau).

+++ js/2_9/main.js
@@ -525,11 +525,13 @@ async function otaUpdate() {
 
   setDisId('otabutton', true);
   try {
-    // 0xA0: bắt đầu — firmware xoá bank không hoạt động
+    // 0xA0: bắt đầu — firmware xoá bank không hoạt động. Size gửi u32
+    // (firmware >64KB làm u16 tràn → xóa thiếu sector → hỏng bank);
+    // firmware cũ đọc u16 vẫn đúng vì 2 byte cao = 0 khi size <64KB.
     const buf = new Uint8Array(136);
     const dv = new DataView(buf.buffer);
     buf[0] = 0xa0; buf[1] = 0x00;
-    dv.setUint16(2, firmSize, true);
+    dv.setUint32(2, firmSize, true);
     show('Đang xoá flash…');
     await write(buf, true);
